third_party/WebKit/LayoutTests/external/wpt/page-visibility/prerender_call-expected.txt

This is a testharness.js-based test.
FAIL VisibilityState of the target page was set to "prerender" when it has been prerenderedassert_equals: visibilityState of the target page was set to 'prerender' expected (string) "hit" but got (object) null
Harness: the test ran to completion.
